+++ src/app/panel/manufacturers/show-manufacturers/show-manufacturers.component.ts
@@ -3,6 +3,7 @@ import { Router } from '@angular/router';
 import { Manufacturer } from 'src/app/models/manufacturer.model';
 import { ManufacturerService } from 'src/app/services/manufacturer.service';
 
+import Swal from 'sweetalert2'
 import SwiperCore, { Navigation, Pagination, Scrollbar, A11y } from 'swiper';
 
 SwiperCore.use([Navigation, Pagination, Scrollbar, A11y]);
@@ -49,7 +50,25 @@ export class ShowManufacturersComponent implements OnInit {
   }
 
   removeManufacturer(id: string) {
-    this.manufacturerService.deleteManufacturer(id);
+    Swal.fire({
+      title: 'Are you sure?',
+      text: "You won't be able to revert this!",
+      icon: 'warning',
+      showCancelButton: true,
+      background: '#CCE2ED',
+      confirmButtonColor: '#56B38F',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, delete it!'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.manufacturerService.deleteManufacturer(id);
+        Swal.fire(
+          'Deleted!',
+          'Your file has been deleted.',
+          'success'
+        )
+      }
+    })
   }
 
   calcElemntWidth() {
